chore: finish renaming component
