+++ frontend/src/app/components/containers/complaints/complaint-filter.tsx
@@ -28,7 +28,7 @@ import UserService from "@service/user-service";
 import { Roles } from "@apptypes/app/roles";
 import { FilterDate } from "@components/common/filter-date";
 import { ValidationMultiSelect } from "@common/validation-multiselect";
-import { ParkSelect } from "@/app/components/common/park";
+import { ParkSelect } from "@/app/components/common/park-select";
 import { OUTCOMES_REQUIRING_ACTIONED_BY } from "@/app/constants/outcomes-requiring-actioned-by";
 
 type Props = {
