+++ frontend/src/app/components/containers/complaints/details/complaint-details-edit.tsx
@@ -79,7 +79,7 @@ import { getCaseFile } from "@/app/store/reducers/case-thunks";
 import { GIROutcomeReport } from "@/app/components/containers/complaints/outcomes/gir-outcome-report";
 import { RootState } from "@/app/store/store";
 import { Roles } from "@/app/types/app/roles";
-import { ParkSelect } from "@/app/components/common/park";
+import { ParkSelect } from "@/app/components/common/park-select";
 import { MapElement, MapObjectType } from "@/app/types/maps/map-element";
 import { selectEquipment } from "@/app/store/reducers/case-selectors";
